refactor: feature based topology
