+++ src/WebClient/app/graveyard/corpses.json
