+++ src/WebClient/app/layout/Footer.tsx
@@ -13,8 +13,8 @@ import {
 } from '@fluentui/react-components';
 import { ArrowDown32Regular } from '@fluentui/react-icons';
 
-import Socials from '@microsoftgraveyard/components/layout/Socials';
-import useDarkMode from '@microsoftgraveyard/hooks/useDarkMode';
+import Socials from '@microsoftgraveyard/layout/Socials';
+import { useDarkMode } from '@microsoftgraveyard/theme';
 
 const useStyles = makeStyles({
   footer: {
