+++ src/WebClient/app/layout/Socials.tsx
@@ -2,7 +2,7 @@ import { JSX } from 'react';
 
 import { Button, Image, makeStyles, tokens } from '@fluentui/react-components';
 
-import useDarkMode from '@microsoftgraveyard/hooks/useDarkMode';
+import { useDarkMode } from '@microsoftgraveyard/theme';
 
 interface ISocial {
   href: string;
