+++ src/WebClient/app/theme/index.ts
@@ -1,2 +1,4 @@
 export { default as DarkModeProvider } from './DarkMode';
 export { default as ThemeProvider } from './Theme';
+
+export { default as useDarkMode } from './useDarkMode';
