+++ src/WebClient/app/theme/useDarkMode.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { DarkModeContext } from '@microsoftgraveyard/providers/DarkMode';
+import { DarkModeContext } from '@microsoftgraveyard/theme/DarkMode';
 
 const useDarkMode = () => useContext(DarkModeContext);
